requirements.txt
fpdf2
requests
python-dotenv
streamlit
anthropic
